Обновление композера Правки запросов, багов Генеральные тесты

+++ controller/comment.php
@@ -43,26 +43,38 @@ $bot->command('delete', function ($message) use ($bot) {
     $text = str_replace('@SitesManagerBot ', '', $text);
     $id = str_replace('/delete -:', '', $text);
 
-    // обращаемся к бд, получаем запись
+    // обращаемся к бд, получаем запись по id
     $db = new CommentRecord();
-    $comment = $db->getComment($id);
 
-    // стучимся в первоисточник и удаляем комментарий там
-    $client = new Client();
-    $response = $client->request('POST', $comment->getDeleteUrl(), [
-        'body' => [
-            'pass' => PASS,
-            'id' => $comment->getCommentId()
-        ]
-    ]);
-    Logger::Debug($response);
-
-    // удаляем запись в бд
-    if ($db->deleteComment($id))
+    if ($comment = $db->getComment($id))
     {
-        // отправим сообщение об успешном удалении
-        $lang = Lang::IncludeFile('command.php', $message->getFrom()->getLanguageCode());
-        $bot->sendMessage($message->getChat()->getId(), $lang['delete']);
+        // если запись есть в БД, то стучимся в первоисточник и удаляем комментарий там
+        $client = new Client();
+        $response = $client->post($comment->getDeleteUrl(), [
+            'form_params' => [
+                'pass' => PASS,
+                'id' => $comment->getCommentId()
+            ]
+        ]);
+
+        // форматируем ответ
+        $response = json_decode($response->getBody()->getContents(), true);
+
+        if ($response == 'ok')
+        {
+            // если запись успешно удалена в первоисточнике, удаляем в локальной БД
+            $db->deleteComment($id);
+
+            // отправим сообщение об успешном удалении
+            $lang = Lang::IncludeFile('command.php', $message->getFrom()->getLanguageCode());
+            $bot->sendMessage($message->getChat()->getId(), $lang['delete']);
+        }
+        else
+        {
+            // отправим сообщение об ошибке
+            $lang = Lang::IncludeFile('other.php', $message->getFrom()->getLanguageCode());
+            $bot->sendMessage($message->getChat()->getId(), $lang['error_comment']);
+        }
     }
     else
     {
@@ -84,7 +96,7 @@ $bot->on(function (Update $update) use ($bot) {
     // если есть комментарии в бд, ожидающие ответа
     if ($comment = $db->checkComment())
     {
-        // получаем данные от телеграмма
+        // получаем данные сообщения
         $message = $update->getMessage();
 
         // подключение языкового файла
@@ -96,12 +108,12 @@ $bot->on(function (Update $update) use ($bot) {
             // откатываем статус комментария
             if ($db->updateComment($comment->getId(), 'NEW'))
             {
-                // отправим сообщение об отмене
+                // отправим сообщение об успешной отмене
                 $bot->sendMessage($message->getChat()->getId(), $lang['answer_canceled']);
             }
             else
             {
-                // отправим сообщение об ошибке
+                // иначе отправим сообщение об ошибке
                 $bot->sendMessage($message->getChat()->getId(), $lang['error_comment']);
             }
         }
@@ -109,8 +121,8 @@ $bot->on(function (Update $update) use ($bot) {
         {
             // стучимся в первоисточник и отправляем ответ туда
             $client = new Client();
-            $response = $client->request('POST', $comment->getAnswerUrl(), [
-                'body' => [
+            $response = $client->post($comment->getAnswerUrl(), [
+                'form_params' => [
                     'pass' => PASS,
                     'ReviewForm' => [
                         'parent' => $comment->getCommentId(),
@@ -118,13 +130,23 @@ $bot->on(function (Update $update) use ($bot) {
                     ]
                 ]
             ]);
-            Logger::Debug($response);
 
-            // удалить запись из БД
-            $db->deleteComment($comment->getId());
+            // форматируем ответ
+            $response = json_decode($response->getBody()->getContents(), true);
 
-            // отправим сообщение об успешном ответе
-            $bot->sendMessage($message->getChat()->getId(), $lang['success_comment']);
+            if ($response == 'ok')
+            {
+                // если ответ успешно доставлен в первоисточник, то удалить запись из локальной БД
+                $db->deleteComment($comment->getId());
+
+                // отправим сообщение об успешном ответе
+                $bot->sendMessage($message->getChat()->getId(), $lang['success_comment']);
+            }
+            else
+            {
+                // отправим сообщение об ошибке
+                $bot->sendMessage($message->getChat()->getId(), $lang['error_comment']);
+            }
         }
     }
